pyproject.toml
[project]
name = "gdrive-mcp-server"
version = "0.1.0"
description = "MCP server for interacting with Google Drive"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.177.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "mcp[cli]>=1.12.2",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[project.scripts]
gdrive-mcp-server = "gdrive_mcp_server:main"
gdrive-test-server = "gdrive_mcp_server.test_server:main"
